Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main(){
-    int a,b;
+    float a,b;
     b = a * a * 3.14;
-    scanf("%d",&a);
+    scanf("%f",&a);
     printf("Area: %d",b);
     return 0;
 }
